Extract src/projmenu.ts — the project context menu and its colour panel
Both in one module because they are one interaction, not two that happen to
sit next to each other: the appearance panel opens either standalone at the
cursor (a colour dot) or as the menu's Appearance submenu, closeColorPop
clears the menu's .sub-open row, and every button in the panel commits and
so closes the whole stack. Splitting them would have each importing the
other's close function — the same argument that kept the reactor badge in
./footer.

Moved whole (one guarded range, 193 lines): SWATCHES, popKey, normalizeHex,
placePop, openColorPopover, closeColorPop, applyColor, setColor, commitHex
and the two $("colorPop") handlers; then ctxKey, projName, FILE_MANAGER,
CtxRow/ctxRowHtml, openCtxMenu, closeCtxMenu, ctxMenuOpen, openTerminalIn,
copyPath, openAppearanceSub and the ctxMenu mouseover/click plus the
document-level contextmenu handler that opens it.

Six callees in main.ts (renderAll, requestLaunch, launchShell,
openProjectFolder, addProjectPath, removeFavorite) — past the ~4 where
per-callee setters stop reading as anything but noise, so one host object,
same as settings.ts and palui.ts. main.ts keeps openColorPopover (the dot
click) and both close functions (its document handlers and Escape); placePop
and openCtxMenu are module-private now that their only callers moved.

Verified mechanical: the relocated block diffed against
`git show HEAD:src/main.ts` differs only by four `export` keywords and six
`host.` prefixes. Caught one false positive from the scripted rewrite while
doing it — the marker it uses to skip the host block was absent, so a
docblock sentence mentioning renderAll() got rewritten too; fixed, and the
script now refuses to run without the marker rather than silently treating
the whole file as body.

No tests: render module, per the Phase-1 rule. main.ts 1,910 -> 1,724.
Gates: tsc --noEmit clean, 368 vitest pass. No Rust touched.
